+++ R/debug_pipeline.R
@@ -1,18 +1,42 @@
+#' Utility to grab pipeline expression from wherever available
+#'
+#' @param call An expression of the parent call, as captured by [`match.call`]
+#' @inheritParams debug_pipeline
+get_pipeline <- function(pipeline, call){
+    if (missing(pipeline) && length(rlang::call_args(call)) == 0) {
+        pipeline <- rstudioapi::getSourceEditorContext()$selection[[1]]$text
+        if (nchar(pipeline) == 0) {
+            stop("Input unavailable. Did you highlight a pipeline?", call. = FALSE)
+        }
+        pipeline <- rlang::parse_expr(pipeline)
+        attr(pipeline, "rstudio") <- TRUE    # flag for replacement of source
+    } else if (all(rlang::call_args_names(call) != "pipeline")) {
+        # handle assignment with `=`
+        call_names <- rlang::call_args_names(call)
+        call_name <- call_names[call_names != "names"]
+        pipeline <- rlang::call2("=", as.name(call_name), call[[call_name]])
+    } else if (is.character(pipeline)) {
+        pipeline <- paste(pipeline, collapse = "")
+        if (nchar(pipeline) == 0) {
+            stop("Input unavailable. Did you highlight a pipeline?", call. = FALSE)
+        }
+        pipeline <- rlang::parse_expr(pipeline)
+    } else if (!is.language(pipeline) && is.language(call$pipeline)) {
+        pipeline <- call$pipeline
+    }
+    pipeline
+}
+
 #' Split a pipeline expression into a list of expressions
 #'
-#' `split_pipeline` takes an unevaluated expression of a pipeline and splits it
-#' into a list of expressions.
-#'
-#' Here "expression" is use colloquially to refer to a bare language object,
-#' i.e. like those returned by [`quote`] or [`rlang::expr`], _not_ an object of
-#' class "expression" like those returned by [`expression`]. The latter can be
-#' turned into the former by subsetting, e.g. `expression(1:5 %>% rev())[[1]]`.
+#' `split_pipeline` takes an expression of a pipeline and splits it into a list
+#' of unevaluated calls.
 #'
 #' `split_pipeline` does not split nested pipelines.
 #'
-#' @param pipeline An unevaluated expression of a pipeline to be split.
+#' @inheritParams debug_pipeline
 #'
-#' @return A list of unevaluated expressions of each element piped together.
+#' @return A list of unevaluated calls extracted from the pipeline.
 #'
 #' @examples
 #' \dontrun{
@@ -24,12 +48,14 @@
 #' }
 #'
 #' @seealso The unexported [\code{magrittr:::split_chain}](https://github.com/tidyverse/magrittr/blob/master/R/split_chain.R),
-#'     which does not handle assignment, but collects the particular pipes used.
+#'     does not handle assignment, but collects the particular pipes used.
+#'     [`burst_pipes`] invisibly returns a restructured split pipeline.
 #' @export
-split_pipeline <- function(pipeline){
+split_pipeline <- function(pipeline, ...){
+    pipeline <- get_pipeline(pipeline, match.call())
     if (inherits(pipeline, "call") && rlang::call_name(pipeline) == "%>%") {
         c(split_pipeline(pipeline[[2]]), pipeline[[3]])
-    } else if (inherits(pipeline, "<-")) {
+    } else if (inherits(pipeline, c("<-", "="))) {
         x <- split_pipeline(pipeline[[3]])
         c(rlang::call_modify(pipeline[1:2], x[[1]]), x[-1])
     } else {
@@ -37,6 +63,166 @@ split_pipeline <- function(pipeline){
     }
 }
 
+#' Recursive helper function for `replace_dots`. Collapses nested nodelist back
+#' into a call.
+#'
+#' @param nodelist A list of calls, with parameters nested, of arbitrary depth
+#'
+#' @seealso [`replace_dots`]
+collapse_nodelist <- function(nodelist) {
+    if (length(nodelist) == 1) return(nodelist)    # return atomics
+    if (any(lengths(as.list(nodelist)) != 1)) {
+        nodelist <- lapply(nodelist, collapse_nodelist)    # recurse nested calls
+    }
+    as.call(nodelist)
+}
+
+#' Recursive helper function that replaces `.` within a call with a replacement
+#' expression. Ignores formulas.
+#'
+#' @return A call with the dots replaced
+#'
+#' @param call A quoted call in which to replace dots
+#' @param replacement A quoted name with which to replace dots
+#'
+#' @seealso [`collapse_nodelist`]
+replace_dots <- function(call, replacement){
+    if (!rlang::is_formula(call) && length(call) > 1) {
+        call <- lapply(call, replace_dots, replacement = replacement)    # recurse
+    } else if (call == as.name(".")) {
+        call <- replacement    # replace
+    }
+    collapse_nodelist(call)
+}
+
+#' Pipeline parser helper function that returns a list of names and calls
+#'
+#' @return A list with `names` and `calls` elements. The `names` element
+#'     contains `data` and `assign` sub-elements.
+#'
+#' @inheritParams burst_pipes
+#'
+#' @seealso [`split_pipeline`], [`burst_pipes`]
+parse_pipeline <- function(pipeline, names){
+    pipe_lines <- split_pipeline(pipeline)
+    if (missing(names)) {
+        names <- paste0("dot", seq_along(pipe_lines[-1]))
+    }
+    assign_names <- lapply(names, as.name)
+    if (inherits(pipe_lines[[1]], c("<-", "="))) {
+        # with assignment
+        data_names <- c(pipe_lines[[1]][[3]], assign_names[-length(assign_names)])
+        assign_names[[length(assign_names)]] <- pipe_lines[[1]][[2]]
+    } else {
+        # no assignment in pipeline
+        data_names <- c(pipe_lines[[1]], assign_names[-length(assign_names)])
+    }
+    pipe_lines[[1]] <- NULL    # drop input data
+    if (length(assign_names) != length(pipe_lines)) {
+        stop("Number of names is not equal to number of top-level calls.",
+             call. = FALSE)
+    }
+
+    # return parsed pipeline object
+    list(
+        names = list(
+            data = data_names,
+            assign = assign_names
+        ),
+        calls = pipe_lines
+    )
+}
+
+#' Replace pipes with intermediate assignment
+#'
+#' `burst_pipes` rearranges a magrittr pipeline into equivalent unpiped code.
+#' Called directly, it will print the restructured code to the console. Called
+#' via the "Burst pipes" RStudio add-in while a pipeline is highlighted, it
+#' will replace the highlighted code with the restructured code.
+#'
+#' Note that nested pipelines are currently ignored. Calling on pipelines from
+#' the inside out should still allow restructuring.
+#'
+#' Warning: Calling `burst_pipes` from the RStudio add-in is currently
+#' irreversible aside from "Edit > Undo" and version control. Save code
+#' beforehand and check equivalence afterwards.
+#'
+#' @return A list of expressions of the restructured, unpiped code, invisibly.
+#'
+#' @param names A character vector of names to be used for intermediary
+#'     assignment of a length equal to that of the number of top-level calls in
+#'     the pipeline.
+#' @inheritParams debug_pipeline
+#'
+#' @seealso [`split_pipeline`] splits a pipeline without restructuring.
+#'
+#' @examples
+#' \dontrun{
+#' library(magrittr)
+#'
+#' burst_pipes(x <- 1:5 %>% rev %>% {. * 2} %>% .[3] %>% rnorm(1, ., sd = ./10))
+#'
+#' burst_pipes(
+#'     x = 1:5 %>% rev %>% {. * 2} %>% .[3] %>% rnorm(1, ., sd = ./10),
+#'     names = c("reversed", "doubled", "x")
+#' )
+#' }
+#'
+#' @export
+burst_pipes <- function(pipeline, names, ...){
+    pipeline <- get_pipeline(pipeline, call = match.call())
+    parts <- parse_pipeline(pipeline, names)
+
+    # Insert first-level data
+    parts$calls <- Map(
+        function(part, name){
+            if (!inherits(part, "{") &&
+                !any(vapply(part, `==`, logical(1), as.name(".")))) {
+                if (is.symbol(part)) {
+                    part <- rlang::call2(part)    # handle parentheses-less calls
+                }
+                rlang::expr(rlang::UQ(part[[1]])(    # function name
+                    !!name,    # data piped in
+                    !!!rlang::call_args(part))    # function parameters
+                )
+            } else {
+                part    # data will be inserted as dot
+            }
+        },
+        parts$calls,
+        parts$names$data
+    )
+
+    # Insert data at dots
+    parts$calls <- Map(
+        replace_dots,
+        call = parts$calls,
+        replacement = parts$names$data
+    )
+
+    # Insert assignment
+    parts$calls <- Map(
+        function(part, name){
+            substitute(n <- p, list(p = part, n = name))
+        },
+        parts$calls,
+        parts$names$assign)
+
+    # Visible returns
+    parts$text <- paste(lapply(parts$calls, function(part){
+        paste(trimws(deparse(part)), collapse = "")
+    }), collapse = "\n")
+    if (isTRUE(attr(pipeline, 'rstudio'))) {
+        rstudioapi::modifyRange(
+            rstudioapi::getSourceEditorContext()$selection[[1]]$range,
+            parts$text
+        )
+    } else {
+        cat(parts$text)
+    }
+
+    invisible(parts$calls)
+}
 
 #' Run a pipeline step-by-step in the debugging browser
 #'
@@ -54,13 +240,11 @@ split_pipeline <- function(pipeline){
 #' For a full explanation of the special commands available in the debugging
 #' browser, see [`browser`].
 #'
-#' @param pipeline A pipeline to debug step-by-step Can be an unquoted
-#'     expression or a character vector of code. If missing, uses the text
-#'     highlighted in RStudio's source editor.
-#' @param data Determines whether data is piped in (`data %>%
-#'     function(parameter)`) or inserted as the first parameter of each call
-#'     (`function(data, parameter)`). Piping is necessary to make dot notation
-#'     work; inserting makes stepping into calls easier.
+#' @param pipeline A pipeline. Can be an unquoted or quoted expression, or a
+#'     character vector of code. If missing, uses the text highlighted in
+#'     RStudio's source editor.
+#' @param ... Only used to allow assignment with `=` at the beginning of a
+#'     pipeline.
 #'
 #' @examples
 #' \dontrun{
@@ -69,59 +253,21 @@ split_pipeline <- function(pipeline){
 #' debug_pipeline(
 #'     x <- 1:5 %>% rev %>% {. * 2} %>% sample(replace = TRUE)
 #' )
-#'
-#' debug_pipeline(
-#'     x <- 1:5 %>% rev() %>% sample(replace = TRUE),
-#'     data = "insert"
-#' )
 #' }
 #'
-#' @seealso [`browser`], [`magrittr::debug_pipe`]
+#' @seealso [`browser`], [`magrittr::debug_pipe`], [burst_pipes]
 #' @export
-debug_pipeline <- function(pipeline, data = c("pipe", "insert")){
-    this_call <- match.call()
-    data = match.arg(data)
+debug_pipeline <- function(pipeline, ...){
+    pipeline <- get_pipeline(pipeline, call = match.call())
 
-    if (missing(pipeline)) {
-        pipeline <- rstudioapi::getSourceEditorContext()$selection[[1]]$text
-        pipeline <- rlang::parse_expr(pipeline)
-    } else if (is.character(pipeline) && nchar(pipeline) > 0) {
-        pipeline <- paste(pipeline, collapse = "")
-        pipeline <- rlang::parse_expr(pipeline)
-    } else if (is.language(this_call[[2]])) {
-        pipeline <- this_call[[2]]
-    } else {
-        stop("Input unavailable. Did you highlight a pipeline?")
-    }
-
-    pipe_parts <- split_pipeline(pipeline)
-    pipe_parts[[1]] <- rlang::expr(print(dot1 <- !!pipe_parts[[1]]))
-    pipe_parts[-1] <- Map(
-        function(line, i){
-            if (data == "insert") {
-                name0 <- as.name(paste0("dot", i))
-                name1 <- as.name(paste0("dot", i + 1))
-                line <- rlang::expr(rlang::UQ(line[[1]])(    # function name
-                    !!name0,    # data piped in
-                    !!!rlang::call_args(line))    # function parameters
-                )
-                substitute(print(nm1 <- ln),
-                           list(nm1 = name1,
-                                ln = line))
-            } else {
-                substitute(print(nm1 <- nm0 %>% ln),
-                           list(nm0 = as.name(paste0("dot", i)),
-                                nm1 = as.name(paste0("dot", i + 1)),
-                                ln = line))
-            }
-        },
-        line = pipe_parts[-1],
-        i = seq_along(pipe_parts[-1])
-    )
+    pipe_lines <- burst_pipes(pipeline, ...)
+    pipe_lines <- lapply(pipe_lines, function(part){
+        rlang::expr(print(!!part))
+    })
 
     pipeline_function <- rlang::new_function(
         args = NULL,
-        body = rlang::expr({!!!pipe_parts})
+        body = rlang::expr({!!!pipe_lines})
     )
     debugonce(pipeline_function)
     pipeline_function()
@@ -129,9 +275,3 @@ debug_pipeline <- function(pipeline, data = c("pipe", "insert")){
     # so browser exits fully when called by add-in
     if (rstudioapi::isAvailable()) rstudioapi::sendToConsole("")
 }
-
-debug_pipeline_insert = debug_pipeline
-rlang::fn_fmls(debug_pipeline_insert)$data <- "insert"
-
-debug_pipeline_pipe = debug_pipeline
-rlang::fn_fmls(debug_pipeline_pipe)$data <- "pipe"